Game saves properly
Had to remove support from V0 to V1 in order to keep enough memory for the game to still work without crash.

Tried to optimize in other areas as well.

+++ src/island.h
@@ -8,6 +8,6 @@ extern const int FOODPRODUCTIONVALUE;
 
 void initialize_islands(GameData* gamedata);
 void update_islands(GameData* gamedata);
-void initialize_island(GameData* gamedata, int islandnumber, int islandtype, int x, int y, int overridemetal, int overridewood,
-                       int overridestone, int overridefood);
+void initialize_island(GameData* gamedata, int8_t islandnumber, int8_t islandtype, int x, int y, int8_t overridemetal, int8_t overridewood,
+                       int8_t overridestone, int8_t overridefood);
 void updateislandresource(GameData* gamedata, int islandnumber, int res1adjust, int res2adjust, int res3adjust, int res4adjust);
